Rename src/render and src/gui C++ sources according to class names
