+++ src/gui/HelpDialog.h
