+++ src/gui/QtLogger.h
