+++ src/gui/ShaderEditor.h
