+++ src/render/HCloudView.cpp
@@ -1,13 +1,13 @@
 // Copyright 2015, Christopher J. Foster and the other displaz contributors.
 // Use of this code is governed by the BSD-style license found in LICENSE.txt
 
-#include "hcloudview.h"
+#include "HCloudView.h"
 
 #include "hcloud.h"
 #include "ClipBox.h"
 #include "glutil.h"
-#include "qtlogger.h"
-#include "shader.h"
+#include "QtLogger.h"
+#include "Shader.h"
 #include "streampagecache.h"
 #include "util.h"
