+++ src/render/TriMesh.cpp
@@ -1,7 +1,7 @@
 // Copyright 2015, Christopher J. Foster and the other displaz contributors.
 // Use of this code is governed by the BSD-style license found in LICENSE.txt
 
-#include "mesh.h"
+#include "TriMesh.h"
 
 #include <memory>
 
@@ -13,7 +13,7 @@
 
 #include "ply_io.h"
 #include "PolygonBuilder.h"
-#include "qtlogger.h"
+#include "QtLogger.h"
 
 
 //------------------------------------------------------------------------------
